requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.2.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
bleach>=6.1.0
python-magic>=0.4.27
